@@ -13,18 +13,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.0.0",
+    "@angular/animations": "^14.1.0",
     "@angular/cdk": "^13.0.2",
-    "@angular/common": "~13.0.0",
-    "@angular/compiler": "~13.0.0",
-    "@angular/core": "~13.0.0",
+    "@angular/common": "^14.1.0",
+    "@angular/compiler": "^14.1.0",
+    "@angular/core": "^14.1.0",
     "@angular/fire": "^7.2.0",
     "@angular/flex-layout": "^13.0.0-beta.36",
-    "@angular/forms": "~13.0.0",
+    "@angular/forms": "^14.1.0",
     "@angular/material": "^13.0.2",
-    "@angular/platform-browser": "~13.0.0",
-    "@angular/platform-browser-dynamic": "~13.0.0",
-    "@angular/router": "~13.0.0",
+    "@angular/platform-browser": "^14.1.0",
+    "@angular/platform-browser-dynamic": "^14.1.0",
+    "@angular/router": "^14.1.0",
     "dotenv": "^16.0.1",
     "firebase": "^9.4.0",
     "firebase-tools": "^9.23.3",
@@ -41,9 +41,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.0.1",
-    "@angular/cli": "~13.0.1",
-    "@angular/compiler-cli": "~13.0.0",
+    "@angular-devkit/build-angular": "^14.1.0",
+    "@angular/cli": "^14.1.0",
+    "@angular/compiler-cli": "^14.1.0",
     "@types/jasmine": "~3.10.0",
     "@types/lodash": "^4.14.181",
     "@types/node": "^12.11.1",
@@ -54,6 +54,6 @@
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
     "ts-node": "^10.8.0",
-    "typescript": "~4.4.3"
+    "typescript": "~4.7.4"
   }
 }
